+++ src/main/java/com/ing/data/cassandra/jdbc/JdbcFloat.java
@@ -54,7 +54,11 @@ public class JdbcFloat extends AbstractJdbcType<Float> {
     }
 
     public String toString(final Float obj) {
-        return obj.toString();
+        if (obj != null) {
+            return obj.toString();
+        } else {
+            return null;
+        }
     }
 
     public boolean needsQuotes() {